store/tikv: move kv.TxnInfo to store/tikv
